clean up code and drop transaction implementation
